chore: cleanup unused old code
