+++ webui/react/src/pages/TrialDetails/Logs/TrialLogFilters.tsx
@@ -1,18 +1,18 @@
 import { Button, Select } from 'antd';
 import { SelectValue } from 'antd/es/select';
-import dayjs, { Dayjs } from 'dayjs';
-import utc from 'dayjs/plugin/utc';
-import React, { useMemo } from 'react';
+import React, { useCallback, useMemo } from 'react';
 
-import DatePickerFilter from 'components/DatePickerFilter';
-import { LogViewerTimestampFilter } from 'components/LogViewerTimestamp';
 import MultiSelect from 'components/MultiSelect';
-
-dayjs.extend(utc);
 
 const { Option } = Select;
 
-export enum LogLevelFromApi {
+interface Props {
+  onChange?: (filters: Filters) => void;
+  options: Filters;
+  values: Filters;
+}
+
+enum LogLevelFromApi {
   Unspecified = 'LOG_LEVEL_UNSPECIFIED',
   Trace = 'LOG_LEVEL_TRACE',
   Debug = 'LOG_LEVEL_DEBUG',
@@ -22,7 +22,7 @@ export enum LogLevelFromApi {
   Critical = 'LOG_LEVEL_CRITICAL',
 }
 
-export interface TrialLogFiltersInterface extends LogViewerTimestampFilter {
+export interface Filters {
   agentIds?: Array<string>,
   containerIds?: Array<string>,
   levels?: Array<LogLevelFromApi>,
@@ -31,132 +31,59 @@ export interface TrialLogFiltersInterface extends LogViewerTimestampFilter {
   stdtypes?: Array<string>,
 }
 
-export interface LogViewerTimestampFilterComponentProp {
-  filter: TrialLogFiltersInterface;
-  filterOptions: TrialLogFiltersInterface;
-  onChange?: (newFilters: TrialLogFiltersInterface) => void;
-}
-
-const TrialLogFilters: React.FC<LogViewerTimestampFilterComponentProp> =
-  ({ filter, filterOptions, onChange }: LogViewerTimestampFilterComponentProp) => {
-
-    const broadcastChange = (newFilter: TrialLogFiltersInterface) => {
-      if (typeof onChange === 'function') {
-        onChange(newFilter);
-      }
+const TrialLogFilters: React.FC<Props> = ({ onChange, options, values }: Props) => {
+  const selectOptions = useMemo(() => {
+    return {
+      ...options,
+      levels: Object.entries(LogLevelFromApi)
+        .filter(entry => entry[1] !== LogLevelFromApi.Unspecified)
+        .map(([ key, value ]) => ({ label: key, value })),
     };
+  }, [ options ]);
 
-    const onAgentChange = (value: SelectValue) => broadcastChange({
-      ...filter,
-      agentIds: (value as Array<string>).map((item) => String(item)),
+  const handleChange = useCallback((
+    key: keyof Filters,
+    caster: NumberConstructor | StringConstructor,
+  ) => (value: SelectValue) => {
+    onChange?.({
+      ...values,
+      [key]: (value as Array<string>).map(item => caster(item)),
     });
+  }, [ onChange, values ]);
 
-    const onClear = () => broadcastChange({});
+  const handleClear = useCallback(() => onChange?.({}), [ onChange ]);
 
-    const onContainerChange = (value: SelectValue) => broadcastChange({
-      ...filter,
-      containerIds: (value as Array<string>).map((item) => String(item)),
-    });
-
-    const onRankChange = (value: SelectValue) => broadcastChange({
-      ...filter,
-      rankIds: (value as Array<string>).map((item) => Number(item)),
-    });
-
-    const onLevelChange = (value: SelectValue) => broadcastChange({
-      ...filter,
-      levels: (value as Array<string>).map((item) => String(item) as LogLevelFromApi),
-    });
-
-    const onDateChange = (key: string, date: Dayjs|null) => {
-      let dateUtc = null;
-
-      if (date) {
-        // receiving a date with user timezone. need to keep the selected date/time but
-        // set the timezone to UTC.
-        const iso8601StringNoTz = date.format().substr(0, 19);
-        dateUtc = dayjs.utc(iso8601StringNoTz);
-      }
-
-      broadcastChange({
-        ...filter,
-        [key]: dateUtc,
-      });
-    };
-
-    const onAfterDateChange = (date: Dayjs|null) => onDateChange('timestampAfter', date);
-
-    const onBeforeDateChange = (date: Dayjs|null) => onDateChange('timestampBefore', date);
-
-    const logLevelList = useMemo(() => {
-      return Object.entries(LogLevelFromApi)
-        .filter(([ key ]) => key !== 'Unspecified')
-        .map(([ key, value ]) => ({ label: key, value }));
-    }, []);
-
-    return (
-      <>
-        <MultiSelect
-          label="Agents"
-          value={filter.agentIds || []}
-          onChange={onAgentChange}
-        >
-          {(filterOptions?.agentIds || []).map((agentId) => (
-            <Option key={agentId} value={agentId}>
-              {agentId}
-            </Option>
-          ))}
-        </MultiSelect>
-        <MultiSelect
-          label="Containers"
-          value={filter.containerIds || []}
-          onChange={onContainerChange}
-        >
-          {(filterOptions?.containerIds || []).map((containerId) => (
-            <Option key={containerId} value={containerId}>
-              {containerId}
-            </Option>
-          ))}
-        </MultiSelect>
-        <MultiSelect
-          label="Ranks"
-          value={filter.rankIds || []}
-          onChange={onRankChange}
-        >
-          {(filterOptions?.rankIds || []).map((rankId) => (
-            <Option key={rankId} value={rankId}>
-              {rankId}
-            </Option>
-          ))}
-        </MultiSelect>
-        <MultiSelect
-          label="Level"
-          value={filter.levels || []}
-          onChange={onLevelChange}
-        >
-          {logLevelList.map((logLevel) => (
-            <Option key={logLevel.value} value={logLevel.value}>
-              {logLevel.label}
-            </Option>
-          ))}
-        </MultiSelect>
-        <DatePickerFilter
-          label="Start"
-          showNow={false}
-          showTime
-          value={filter.timestampAfter}
-          onChange={onAfterDateChange}
-        />
-        <DatePickerFilter
-          label="End"
-          showNow={false}
-          showTime
-          value={filter.timestampBefore}
-          onChange={onBeforeDateChange}
-        />
-        <Button onClick={onClear}>Clear</Button>
-      </>
-    );
-  };
+  return (
+    <>
+      <MultiSelect
+        label="Agents"
+        value={values.agentIds}
+        onChange={handleChange('agentIds', String)}>
+        {(selectOptions?.agentIds || []).map(id => <Option key={id} value={id}>{id}</Option>)}
+      </MultiSelect>
+      <MultiSelect
+        label="Containers"
+        value={values.containerIds}
+        onChange={handleChange('containerIds', String)}>
+        {(selectOptions?.containerIds || []).map(id => <Option key={id} value={id}>{id}</Option>)}
+      </MultiSelect>
+      <MultiSelect
+        label="Ranks"
+        value={values.rankIds}
+        onChange={handleChange('rankIds', Number)}>
+        {(selectOptions?.rankIds || []).map(id => <Option key={id} value={id}>{id}</Option>)}
+      </MultiSelect>
+      <MultiSelect
+        label="Level"
+        value={values.levels}
+        onChange={handleChange('levels', String)}>
+        {selectOptions?.levels.map((level) => (
+          <Option key={level.value} value={level.value}>{level.label}</Option>
+        ))}
+      </MultiSelect>
+      <Button onClick={handleClear}>Clear</Button>
+    </>
+  );
+};
 
 export default TrialLogFilters;
